fix: drop importing LED settings from a theme
LEDs are usually set manually by users and can even be done on a per-console or game basis, and thus do not really belong in themes. It is removed in this commit to avoid issues where users have to go back and reset their LED settings after adding a theme.